chatwithsite/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.6
crawl4ai>=0.4.247
pydantic>=2.10.0
python-dotenv>=1.0.1
openai>=1.51.0
langchain>=0.3.5
langchain-openai>=0.2.5
langchain-community>=0.3.4
faiss-cpu>=1.9.0
tiktoken>=0.8.0